Updated how settings versions are handled to make more sense
