+++ src/bin/hdr-snipping-tool/settings/pre_1_2_0_settings.rs
@@ -1,0 +1,8 @@
+use livesplit_hotkey::KeyCode;
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+pub struct Pre1_2_0Settings {
+    pub version: String,
+    pub screenshot_key: KeyCode,
+}
